Improved ToCellMapperTest.

+++ FantasyWorldSimulation/test/jfws/util/map/SharedTestData.java
@@ -30,32 +30,32 @@ public class SharedTestData {
 
 	public static final double ERROR = 0.001;
 
-	public static final int ORIGIN_X = 3;
-	public static final int ORIGIN_Y = 4;
-	public static final int RESOLUTION_X = 2;
-	public static final int RESOLUTION_Y = 3;
+	public static final double ORIGIN_X = 3;
+	public static final double ORIGIN_Y = 4;
+	public static final double RESOLUTION_X = 2;
+	public static final double RESOLUTION_Y = 3;
 
-	public static final int MAX_X = ORIGIN_X + WIDTH * RESOLUTION_X - 1;
-	public static final int MAX_Y = ORIGIN_Y + HEIGHT * RESOLUTION_Y - 1;
+	public static final double MAX_X = ORIGIN_X + WIDTH * RESOLUTION_X - 1;
+	public static final double MAX_Y = ORIGIN_Y + HEIGHT * RESOLUTION_Y - 1;
 
-	public static final int CELL_X_0 = ORIGIN_X;
-	public static final int CELL_X_1 = ORIGIN_X + 1 * RESOLUTION_X;
-	public static final int CELL_X_2 = ORIGIN_X + 2 * RESOLUTION_X;
-	public static final int CELL_X_3 = ORIGIN_X + 3 * RESOLUTION_X;
+	public static final double CELL_X_0 = ORIGIN_X;
+	public static final double CELL_X_1 = ORIGIN_X + 1 * RESOLUTION_X;
+	public static final double CELL_X_2 = ORIGIN_X + 2 * RESOLUTION_X;
+	public static final double CELL_X_3 = ORIGIN_X + 3 * RESOLUTION_X;
 
-	public static final int CELL_Y_0 = ORIGIN_Y;
-	public static final int CELL_Y_1 = ORIGIN_Y + 1 * RESOLUTION_Y;
-	public static final int CELL_Y_2 = ORIGIN_Y + 2 * RESOLUTION_Y;
-	public static final int CELL_Y_3 = ORIGIN_Y + 3 * RESOLUTION_Y;
-	public static final int CELL_Y_4 = ORIGIN_Y + 4 * RESOLUTION_Y;
+	public static final double CELL_Y_0 = ORIGIN_Y;
+	public static final double CELL_Y_1 = ORIGIN_Y + 1 * RESOLUTION_Y;
+	public static final double CELL_Y_2 = ORIGIN_Y + 2 * RESOLUTION_Y;
+	public static final double CELL_Y_3 = ORIGIN_Y + 3 * RESOLUTION_Y;
+	public static final double CELL_Y_4 = ORIGIN_Y + 4 * RESOLUTION_Y;
 
-	public static final int CENTER_X_0 = ORIGIN_X + 1 * RESOLUTION_X / 2;
-	public static final int CENTER_X_1 = ORIGIN_X + 3 * RESOLUTION_X / 2;
-	public static final int CENTER_X_2 = ORIGIN_X + 5 * RESOLUTION_X / 2;
+	public static final double CENTER_X_0 = ORIGIN_X + 1 * RESOLUTION_X / 2;
+	public static final double CENTER_X_1 = ORIGIN_X + 3 * RESOLUTION_X / 2;
+	public static final double CENTER_X_2 = ORIGIN_X + 5 * RESOLUTION_X / 2;
 
-	public static final int CENTER_Y_0 = ORIGIN_Y + 1 * RESOLUTION_Y / 2;
-	public static final int CENTER_Y_1 = ORIGIN_Y + 3 * RESOLUTION_Y / 2;
-	public static final int CENTER_Y_2 = ORIGIN_Y + 5 * RESOLUTION_Y / 2;
+	public static final double CENTER_Y_0 = ORIGIN_Y + 1 * RESOLUTION_Y / 2;
+	public static final double CENTER_Y_1 = ORIGIN_Y + 3 * RESOLUTION_Y / 2;
+	public static final double CENTER_Y_2 = ORIGIN_Y + 5 * RESOLUTION_Y / 2;
 
 	public static final ToCellMapper<Integer> MAPPER = new ToCellMapper<>(MAP, ORIGIN_X, ORIGIN_Y, RESOLUTION_X, RESOLUTION_Y);
 }
